requirements.txt
crewai==0.30.11
crewai-tools==0.2.6
cryptography==42.0.8
email_validator==2.1.1
fastapi-cli==0.0.4
flexdown==0.1.5a8
GitPython==3.1.43
googletrans-py==4.0.0
grpcio-status==1.62.2
httptools==0.6.1
install==1.3.5
mistletoe==1.3.0
Pillow==9.5.0
plotly-express==0.4.1
psycopg2-binary==2.9.9
PyGithub==1.59.1
PySocks==1.7.1
reflex-chat==0.0.2a1
reflex-image-zoom==0.0.2
reflex-type-animation==0.0.1
shellingham==1.5.4
typesense==0.14.0
ujson==5.10.0
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
